# backend/openapi-spec-v2.yaml
openapi: 3.1.0
info:
  title: Govpulse Backend Service
  version: 2.0.0
  description: |
    This is the backend service for the Govpulse application, providing endpoints for user management, issue tracking, appointment scheduling, and more.
    With v2 it supports Clerk for user authentication and authorization.
...
paths:
...
  /api/v2/user/signup-via-clerk-token:
    get:
      summary: Clerk supported signup
      operationId: clerkSupportedSignup
      description: Allows a user to sign up using Clerk token. This is a v2 endpoint, not available in earlier versions.
      tags:
        - user-related
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "User registered successfully through backend"

        "400":
          description: Bad Request - Invalid input data
          content:
            text/plain:
              schema:
                type: string
                example: "Bad Request - Invalid input data"
        "401":
          description: Unauthorized - Invalid token or insufficient permissions
          content:
            text/plain:
              schema:
                type: string
                example: "Unauthorized - Invalid token or insufficient permissions"
        "500":
          description: Internal Server Error
          content:
            text/plain:
              schema:
                type: string
                example: "Internal Server Error"
      deprecated: false
      security: []
  /api/v2/issues/create:
    post:
      summary: Create a new issue
      operationId: createIssue
      description: Allows a user to create a new issue
      tags:
        - issues-related
      requestBody:
        description: Contains the details of the issue to be created
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/models.issueCreateReqObj"
      responses:  
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/models.issueCreateResObj"
        "400":
          description: Bad Request - Invalid input data
          content:
            text/plain:
              schema:
                type: string
                example: Bad Request - Invalid input data
        "401":
          description: Unauthorized - Invalid token
          content:
            text/plain:
              schema:
                type: string
                example: Unauthorized - Invalid token              
        "500":
          description: Internal Server Error
          content:
            text/plain:
              schema:
                type: string
                example: Internal Server Error
      deprecated: false
      security: 
        - BearerAuth: []
  /api/v2/issues/delete-issue/{issue_id}:
    delete:
      summary: Delete an issue
      operationId: deleteIssue
      description: Allows a user to delete an issue by its ID
      tags:
        - issues-related
      parameters:
        - name: issue_id
          in: path
          required: true
          description: The ID of the issue to be deleted
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:  
                    type: string
                    example: Issue deleted successfully
        "401":
          description: Unauthorized - Invalid token
          content:
            text/plain:
              schema:
                type: string
                example: Unauthorized - Invalid token
        "404":
          description: Not Found - Issue not found
          content:
            text/plain:
              schema:
                type: string
                example: Not Found - Issue not found
        "500":
          description: Internal Server Error
          content:
            text/plain:
              schema:
                type: string
                example: Internal Server Error
      deprecated: false
      security: 
        - BearerAuth: []
  /api/v2/issues/user-issues:
    get:
      summary: Get issues created by the user
      operationId: getUserIssues
      description: Retrieves all issues created by the authenticated user
      tags:
        - issues-related
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/models.issuedata"
        "500":
          description: Internal Server Error
          content:
            text/plain:
              schema:
                type: string
                example: Internal Server Error
      deprecated: false
      security: 
        - BearerAuth: []
  /api/v2/issues/issues:
    get:
      summary: Get all issues
      operationId: getIssues
      description: Retrieves all issues in the system
      tags:
        - issues-related
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                    $ref: "#/components/schemas/models.issuedata"
        "500":
          description: Internal Server Error
          content:
            text/plain:
              schema:
                type: string
                example: Internal Server Error
      deprecated: false
      security: []
  /api/v2/issues/issues/{issue_id}:
    get:
      summary: Get issue by ID
      operationId: getIssueById
      description: Retrieves a specific issue by its ID
      tags:
        - issues-related
      parameters:
        - name: issue_id
          in: path
          required: true
          description: The ID of the issue to retrieve
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/models.issuedata"
        "404":
          description: Not Found - Issue not found
          content:
            text/plain:
              schema:
                type: string
                example: Not Found - Issue not found
        "500":
          description: Internal Server Error
          content:
            text/plain:
              schema:
                type: string
                example: Internal Server Error
      deprecated: false
      security: []
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    models.issueCreateReqObj:
      type: object
      properties:
        title:
          type: string
          example: "Issue Title"
        description:
          type: string
          example: "Detailed description of the issue."
        gs_division:
          type: string
          example: "Colombo North"
        ds_division:
          type: string
          example: "Colombo"
        urgency_score:
          type: number
          format: float
          example: 7.5
        status_id:
          type: integer
          format: int64
          example: 1
        authority_id:
          type: integer
          format: int64
          example: 2
        category_id:
          type: integer
          format: int64
          example: 3
        image_urls:
          type: array
          items:
            type: string
            format: uri
            example:
            - "https://example.com/image1.jpg"
            - "https://example.com/image2.jpg"
      required:
        - title
        - description
        - urgency_score
        - authority_id
        - category_id
    models.issueCreateResObj:
      type: object
      properties:
        message:
          type: string
          example: "Issue created successfully"
        issueId:
          type: integer
          format: int64
          example: 12345
    models.issuedata:
      type: object
      properties:
        issue_id:
          type: integer
          format: int64
          example: 6
        user_id:
          type: integer
          format: int64
          example: 6
        title:
          type: string
          example: "test issue3"
        description:
          type: string
          example: "something happened."
        district:
          type: string
          nullable: true
          example: null
        gs_division:
          type: string
          nullable: true
          example: null
        ds_division:
          type: string
          nullable: true
          example: null
        urgency_score:
          type: number
          format: float
          nullable: true
          example: null
        created_at:
          type: string
          format: date-time
          example: "2025-08-11T06:19:58.757Z"
        status_id:
          type: integer
          format: int64
          example: 1
        authority_id:
          type: integer
          format: int64
          nullable: true
          example: null
        category_id:
          type: integer
          format: int64
          nullable: true
          example: null
        image_urls:
          type: array
          items:
            type: string
            format: uri
          nullable: true
          example: null
        approved_for_appointment_placing:
          type: boolean
          example: false
        User:
          type: object
          properties:
            user_id:
              type: integer
              format: int64
              example: 6
            first_name:
              type: string
              example: "test"
            last_name:
              type: string
              example: "user"
            name:
              type: string
              nullable: true
              example: null
            email:
              type: string
              format: email
              example: "testuser1@gmail.com"
            password:
              type: string
              example: "$2b$10$/X9VWOEwy1.e6M8CwMr7EuG/9j.cI6ku2H978Gea3p4dfsGzuWk7O"
            nic:
              type: string
              example: "199902400427"
            role:
              type: string
              example: "citizen"
            profile_image_url:
              type: string
              format: uri
              nullable: true
              example: null
        Issue_Status:
          type: object
          properties:
            status_id:
              type: integer
              format: int64
              example: 1
            status_name:
              type: string
              example: "Pending Review"
            authority_id:
              type: integer
              format: int64
              example: 1
...